Add thin tokens lexer
Add a curosr that operates directly on `&str` and produces
simple(thin) tokens which are a pair of kind and length of
the lexeme. At this step, errors are not reported, but are
instead stored as flags on the tokens. Tokens at this step
are not yet ready for parsing BetterC syntax, as they will
first have to be turned into rich(wide) tokens used by the
actual parser.