Add a basic CLI and lay the base for the language syntax
Add a CLI that takes an argument for the path of the source
and outputs debug information about what lexemes the lexer
generated as well as the statements generated by the parser
from the aforementioned lexemes.

Add a syntax module which allows lexing a source file as
well as turning the output of the lexer into statements.
Right now only structs containing an identifier, generic
type arguments and struct fields are recognised patterns.
Struct fields are pairs of an identifier and a type (which
right now is just an identifier).